Overhaul app.py for the updated API
Diego and I discussed a much richer API for reporting information about a plan and providing that for search interfaces. This requires a significantly different data structure, and so much more interesting processing. This checkin gets most of that done, but leaves out information about actual experiences because that is much more complicated to assemble.